Now stores values in Cheapest Time
Now stores values entered in the "Power" and the "Total consumption" field to persistent storage. The next time the user enteres some data in those fields, closes the app completely and reopens it, the app will read the stored values and pre-enter them. This allows a more faster usage of the app for the user.

+++ App/AWattPrice/AWattPrice/Data/CheapestHourManager.swift
@@ -128,12 +128,12 @@ class CheapestHourManager: ObservableObject {
     
     func calculateHourlyPrice(cheapestHourPair: HourPair, currentSetting: CurrentSetting) -> HourPair {
         if currentSetting.setting!.awattarTariffIndex == 0 {
-            let electricityPriceNoBonus = Float(cheapestHourPair.associatedPricePoints.count) * currentSetting.setting!.awattarBaseElectricityPrice
+            let electricityPriceNoBonus = Double(cheapestHourPair.associatedPricePoints.count) * currentSetting.setting!.awattarBaseElectricityPrice
             if currentSetting.setting!.pricesWithTaxIncluded {
-                cheapestHourPair.hourlyEnergyCosts = electricityPriceNoBonus + cheapestHourPair.countAllPricesTogether(withVat: true)
+                cheapestHourPair.hourlyEnergyCosts = Float(electricityPriceNoBonus) + cheapestHourPair.countAllPricesTogether(withVat: true)
                 
             } else {
-                cheapestHourPair.hourlyEnergyCosts = electricityPriceNoBonus + cheapestHourPair.countAllPricesTogether(withVat: false)
+                cheapestHourPair.hourlyEnergyCosts = Float(electricityPriceNoBonus) + cheapestHourPair.countAllPricesTogether(withVat: false)
             }
         }
         return cheapestHourPair
